+++ app/forms/export/new_rpa_sug_and_fa_letters_form.rb
@@ -1,0 +1,8 @@
+class Export::NewRpaSugAndFaLettersForm < Export::BaseForm
+  def export
+    projects = Project.conversions.significant_date_in_range(from_date.to_s, to_date.to_s)
+    pre_fetched_projects = AcademiesApiPreFetcherService.new.call!(projects)
+
+    Export::Conversions::SchoolsDueToConvertCsvExportService.new(pre_fetched_projects).call
+  end
+end