+++ app/services/projects_for_export_service.rb
@@ -13,9 +13,4 @@ class ProjectsForExportService
     projects = Transfer::Project.filtered_by_significant_date(month, year)
     AcademiesApiPreFetcherService.new.call!(projects)
   end
-
-  def conversion_projects_by_significant_date(month:, year:)
-    projects = Conversion::Project.filtered_by_significant_date(month, year)
-    AcademiesApiPreFetcherService.new.call!(projects)
-  end
 end